qa: better deprecation of $routeIdentifierPlaceholder and population of $identifiersToPlaceholdersMapping
This patch sets the default value of $identifiersToPlaceholdersMapping to an empty array.
When the constructor determines that the `$resourceIdentifier` is not in the `$identifiersToPlaceholdersMapping`, it adds it, setting the value to what is provided in `$routeIdentifierPlaceholder`.
This largely preserves previous behavior, while allowing the simplified routine in `RouteBasedResourceStrategy` to work.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>