qa: allow willdurand/negotiation 2.3 series
If you are not on PHP 8.0, you can still use willdurand/negotiation 2.3.1+, as we are never directly referencing the `Match` class.
However, we cannot test lowest dependencies with that constraint using PHP 8, as we also have the `--ignore-platform-reqs` flag, which will bring in the older, broken-in-php-8.0 version.
As such, I've also removed the `DEPS=LOWEST` PHP 8.0 job.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>