Fixes #50 - allow embedding objects that are part of a union type in `HalResource`
As discussed with @weierophinney in https://github.com/mezzio/mezzio-hal/issues/50,
it is very normal to have a JSON-Schema like following for an API resource:

```json
{
  "components": {
    "SomeResource": {
      "oneOf": [
        {"$ref": "#/components/Foo"},
        {"$ref": "#/components/Tab"}
      ]
    },
    "Foo": {
      "type": "object",
      "required": ["bar"],
      "properties": {
        "bar": {"type": "string"}
      }
    },
    "Tab": {
      "type": "object",
      "required": ["taz"],
      "properties": {
        "taz": {"type": "string"}
      }
    }
  }
}
```

Attempting to store a `Foo` and a `Tab` inside the same `HalResource` within `mezzio/mezzio-hal`
is currently not supported, since `HalResource` attempts to verify that all keys are the same.

This also breaks when declaring `required: [...]` fields on an `object` type, and then some fields
are missing on some elements (valid, same type, not understood by `mezzio/mezzio-hal`).

With this change, it is instead possible to serialize them: it is still important that each embedded
resource has an identification field that can be used for re-assembling its full resource URI.

Signed-off-by: Marco Pivetta <ocramius+dco-sucks@gmail.com>
Signed-off-by: Marco Pivetta <ocramius@gmail.com>